refactor(attribute): Moved persistense directory into infrastructure
