+++ module/attribute/src/Infrastructure/Persistence/Projector/Attribute/DbalAttributeBoolParameterChangeEventProjector.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types = 1);
+
+namespace Ergonode\Attribute\Infrastructure\Persistence\Projector\Attribute;
+
+use Ergonode\Attribute\Domain\Event\Attribute\AttributeBoolParameterChangeEvent;
+
+/**
+ */
+class DbalAttributeBoolParameterChangeEventProjector extends AbstractDbalAttributeParameterChangeEventProjector
+{
+    /**
+     * @param AttributeBoolParameterChangeEvent $event
+     */
+    public function __invoke(AttributeBoolParameterChangeEvent $event): void
+    {
+        $this->projection($event->getAggregateId(), $event->getName(), $event->getTo());
+    }
+}
