+++ module/attribute/src/Infrastructure/Persistence/Projector/Attribute/DbalAttributeDeletedEventProjector.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types = 1);
+
+namespace Ergonode\Attribute\Infrastructure\Persistence\Projector\Attribute;
+
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\DBALException;
+use Ergonode\Attribute\Domain\Event\Attribute\AttributeDeletedEvent;
+
+/**
+ */
+class DbalAttributeDeletedEventProjector
+{
+    private const TABLE = 'attribute';
+
+    /**
+     * @var Connection
+     */
+    private Connection $connection;
+
+    /**
+     * @param Connection $connection
+     */
+    public function __construct(Connection $connection)
+    {
+        $this->connection = $connection;
+    }
+
+    /**
+     * @param AttributeDeletedEvent $event
+     *
+     * @throws DBALException
+     */
+    public function __invoke(AttributeDeletedEvent $event): void
+    {
+        // @todo What we should do with unused values?
+
+        $this->connection->delete(
+            self::TABLE,
+            [
+                'id' => $event->getAggregateId()->getValue(),
+            ]
+        );
+    }
+}
