+++ module/attribute/src/Infrastructure/Persistence/Projector/Group/DbalAttributeGroupCreatedEventProjector.php
@@ -1,0 +1,59 @@
+<?php
+
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types = 1);
+
+namespace Ergonode\Attribute\Infrastructure\Persistence\Projector\Group;
+
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\DBALException;
+use Ergonode\Attribute\Domain\Event\Group\AttributeGroupCreatedEvent;
+use JMS\Serializer\SerializerInterface;
+
+/**
+ */
+class DbalAttributeGroupCreatedEventProjector
+{
+    private const TABLE = 'attribute_group';
+
+    /**
+     * @var Connection
+     */
+    private Connection $connection;
+
+    /**
+     * @var SerializerInterface
+     */
+    private SerializerInterface $serializer;
+
+    /**
+     * @param Connection          $connection
+     * @param SerializerInterface $serializer
+     */
+    public function __construct(Connection $connection, SerializerInterface $serializer)
+    {
+        $this->connection = $connection;
+        $this->serializer = $serializer;
+    }
+
+    /**
+     * @param AttributeGroupCreatedEvent $event
+     *
+     * @throws DBALException
+     */
+    public function __invoke(AttributeGroupCreatedEvent $event): void
+    {
+        $this->connection->insert(
+            self::TABLE,
+            [
+                'id' => $event->getAggregateId()->getValue(),
+                'code' => $event->getCode(),
+                'name' => $this->serializer->serialize($event->getName(), 'json'),
+            ]
+        );
+    }
+}
