+++ module/attribute/src/Infrastructure/Persistence/Projector/Group/DbalAttributeGroupNameChangedEventProjector.php
@@ -1,0 +1,60 @@
+<?php
+
+/**
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+declare(strict_types = 1);
+
+namespace Ergonode\Attribute\Infrastructure\Persistence\Projector\Group;
+
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\DBALException;
+use Ergonode\Attribute\Domain\Event\Group\AttributeGroupNameChangedEvent;
+use JMS\Serializer\SerializerInterface;
+
+/**
+ */
+class DbalAttributeGroupNameChangedEventProjector
+{
+    private const TABLE = 'attribute_group';
+
+    /**
+     * @var Connection
+     */
+    private Connection $connection;
+
+    /**
+     * @var SerializerInterface
+     */
+    private SerializerInterface $serializer;
+
+    /**
+     * @param Connection          $connection
+     * @param SerializerInterface $serializer
+     */
+    public function __construct(Connection $connection, SerializerInterface $serializer)
+    {
+        $this->connection = $connection;
+        $this->serializer = $serializer;
+    }
+
+    /**
+     * @param AttributeGroupNameChangedEvent $event
+     *
+     * @throws DBALException
+     */
+    public function __invoke(AttributeGroupNameChangedEvent $event): void
+    {
+        $this->connection->update(
+            self::TABLE,
+            [
+                'name' => $this->serializer->serialize($event->getTo(), 'json'),
+            ],
+            [
+                'id' => $event->getAggregateId()->getValue(),
+            ]
+        );
+    }
+}
