--- module/attribute/src/Persistence/Dbal/Projector/Attribute/AttributeDeletedEventProjector.php
@@ -1,51 +1,0 @@
-<?php
-
-/**
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE.txt for license details.
- */
-
-declare(strict_types = 1);
-
-namespace Ergonode\Attribute\Persistence\Dbal\Projector\Attribute;
-
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\DBALException;
-use Ergonode\Attribute\Domain\Event\Attribute\AttributeDeletedEvent;
-
-/**
- */
-class AttributeDeletedEventProjector
-{
-    private const TABLE = 'attribute';
-
-    /**
-     * @var Connection
-     */
-    private Connection $connection;
-
-    /**
-     * @param Connection $connection
-     */
-    public function __construct(Connection $connection)
-    {
-        $this->connection = $connection;
-    }
-
-    /**
-     * @param AttributeDeletedEvent $event
-     *
-     * @throws DBALException
-     */
-    public function __invoke(AttributeDeletedEvent $event): void
-    {
-        // @todo What we should do with unused values?
-
-        $this->connection->delete(
-            self::TABLE,
-            [
-                'id' => $event->getAggregateId()->getValue(),
-            ]
-        );
-    }
-}
