--- module/attribute/src/Persistence/Dbal/Projector/Group/AttributeGroupDeletedEventProjector.php
@@ -1,49 +1,0 @@
-<?php
-
-/**
- * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
- * See LICENSE.txt for license details.
- */
-
-declare(strict_types = 1);
-
-namespace Ergonode\Attribute\Persistence\Dbal\Projector\Group;
-
-use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\DBALException;
-use Ergonode\Attribute\Domain\Event\Group\AttributeGroupDeletedEvent;
-
-/**
- */
-class AttributeGroupDeletedEventProjector
-{
-    private const TABLE = 'attribute_group';
-
-    /**
-     * @var Connection
-     */
-    private Connection $connection;
-
-    /**
-     * @param Connection $connection
-     */
-    public function __construct(Connection $connection)
-    {
-        $this->connection = $connection;
-    }
-
-    /**
-     * @param AttributeGroupDeletedEvent $event
-     *
-     * @throws DBALException
-     */
-    public function __invoke(AttributeGroupDeletedEvent $event): void
-    {
-        $this->connection->delete(
-            self::TABLE,
-            [
-                'id' => $event->getAggregateId()->getValue(),
-            ]
-        );
-    }
-}
